# doConcurrentLayoutAndUpdate to speedup Score::update() (Not a PR...just sharing for thoughts)