--- a/mtest/testutils.cpp
+++ b/mtest/testutils.cpp
@@ -41,6 +41,7 @@ inline void initMyResources() {
 extern Ms::Score::FileError importOve(Ms::MasterScore*, const QString& name);
 
 Q_LOGGING_CATEGORY(undoRedo, "undoRedo", QtCriticalMsg)
+Q_LOGGING_CATEGORY(layout, "layout", QtCriticalMsg)
 // Q_LOGGING_CATEGORY(undoRedo, "undoRedo", QtDebugMsg)
 
 namespace Ms {