@ericfont 2019-01-03
(note...travis mtest build will fail, because I have to do something to make it build properly on linux.