@ericfont 2019-01-03
As I was mentioning, initBeamMetrics (https://github.com/musescore/MuseScore/blob/26d15105ab26c47c5866621051d6b025d4a28d36/libmscore/beam.cpp#L676) causes crash due to race condition on first load...so I think if I moved that init stuff  outside of layout call path will work.  FYI: here is the stack trace from that crash: https://gist.github.com/ericfont/681865eeba1ab2d0e06d0bb5e5b5bc8b