@ericfont 2019-01-03
Just some more stats running on my 4-core x86-64 linux machine now (debug build, unfortunately I haven't build release):
- Initial full layout & update: 10627 ms single-threaded gets sped up to 5102 ms...about twice as fast 
- Changing one note: 2855 ms single-threaded becomes 2035...about %40 speedup