@ericfont 2019-01-03
Ok, I think I fixed the initBeamMetrics race conditions using a simple QMutex with a tryLock.

And I think I fixed the Travis linux compiler error by converting qCDebug statements from being like printf statements into being like C++ cout-style statements (I don't know what that works...according to Qt docs, both styles work).