@ericfont 2019-01-03
I can't figure out why Travis is unable to build with the qCDebug statements, either in printf format or in regular format...they're perfectly valid https://doc.qt.io/qt-5/qloggingcategory.html#qCDebug and I defined the logging categories.  It builds fine & passes tests on both my windows MSVC and linux QtCreator machines.