@Justin135 2019-01-03
Maybe you could also add a simple script that updates only the parts that have the edited instrument. So, instead of updating all of the parts in the background, only the one part needs to be updated.