@Justin135 2019-01-04
I would help editing the code, but I'm not familiar with C++. I do know Java, and I've been thinking about learning C++, so I feel like it will make learning the syntax easier.