@ericfont 2019-01-04
@Justin135, I'm not going got try to make too many changes in the PR because I'm keeping it focused just on making multithreaded.

the 3.0 code already does try to do a related optimization in this if block for "update": https://github.com/musescore/MuseScore/blob/e2950adf4ce07cfb8291b6404967ca9766455dd1/libmscore/cmd.cpp#L235.  But it should be noted that the term "update" here has a specific meaning...it means "redraw".  The code tries to be smart by only doing redraws for the current screen for the refresh rectangle of the part that changed.  But it doesn't try limiting the relayout in the main score to be only those instruments.

I've also been noticing that a lot of Score::setLayoutAll() calls seem excessive & unnecessary to me (https://musescore.org/en/node/281091), but again that is a separate issue.

But really quite a lot of changes will have to be made to do what you said.  First the Score::setLayout(int tick) function will have to have an option to only mark layout as dirty for a specific instrument.  So then would have to keep track of dirty or clean state for every measure for each staff.  And then would have to rework the doLayoutRange code such that it can reuse old clean staves and measures.  That will be quite a more involved fix then merely "add a simple script".