@ericfont 2019-01-04
I'm debugging the crash from the failed test (exchange-voice-part.script) and notice it is because inside Score:: doLayoutRange in call for getNextMeasure i does Chord:: layoutStem1 which has to "Create stem if necessary", and that stem is put onto the *master score's* undostack...  So that breaks the thread safeness.  Maybe putting a QMutex to lock undoAddElement for each score will do the trick.