@ericfont 2019-01-04
I think I got tests working now...turns out that Scores may modify the MasterScore's UndoStack during layout, so in order to allow multiple concurrent scores doing layout, I put a QMutex around every UndoStack function which modified the UndoStack.  If each score had it's own undostack for the purposes of layout, then I wouldn't have to do this.